Based on coordinates(lat, lon) the Web Api returns all the groups around.

+++ Boongaloo/Boongaloo.DTO/BoongalooWebApiDto/GroupDto.cs
@@ -12,8 +12,8 @@ namespace Boongaloo.DTO.BoongalooWebApiDto
         public IEnumerable<TagEnum> Tags { get; set; }
 
         #region These should not be null if NewAreaGroup is true
-        public decimal? Latitutude { get; set; }
-        public decimal? Longtitude { get; set; }
+        public double? Latitutude { get; set; }
+        public double? Longtitude { get; set; }
         public RadiusEnum? Radius { get; set; }
         #endregion
